# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v2/companies": {
      "get": {
        "description": "All companies — mining operators and oil & gas operators in one list — with logos, stock metadata and project counts. Filter by `type` (oil_and_gas | mining | both | all), `q`, `commodity`, `country`.",
        "operationId": "CompaniesController_list_v2",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by company type. `all` (default) returns every company.",
            "schema": {
              "example": "oil_and_gas",
              "type": "string",
              "enum": [
                "oil_and_gas",
                "mining",
                "both",
                "all"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive substring search across company name.",
            "schema": {
              "example": "ypf",
              "type": "string"
            }
          },
          {
            "name": "commodity",
            "required": false,
            "in": "query",
            "description": "Only companies with at least one project of this commodity (case-insensitive).",
            "schema": {
              "example": "Uranio",
              "type": "string"
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "description": "Filter by country (case-insensitive exact match).",
            "schema": {
              "example": "Argentina",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CompanyListItemDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          }
        },
        "summary": "List companies",
        "tags": [
          "companies"
        ]
      }
    },
...
    "/api/v2/companies/{slug}": {
      "get": {
        "description": "Full profile: logo, website, stock data, mineral project portfolio (with timeline) and — for oil & gas operators — a production summary.",
        "operationId": "CompaniesController_detail_v2",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "ypf",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CompanyDetailDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Company slug not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Company detail",
        "tags": [
          "companies"
        ]
      }
    },
    "/api/v2/provinces": {
      "get": {
        "description": "Every province with oil & gas or mining activity, with combined project counts and commodities. Filter by `commodity`.",
        "operationId": "ProvincesController_list_v2",
        "parameters": [
          {
            "name": "commodity",
            "required": false,
            "in": "query",
            "description": "Only provinces with at least one mineral project of this commodity (case-insensitive).",
            "schema": {
              "example": "Uranio",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProvinceListItemDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          }
        },
        "summary": "List provinces",
        "tags": [
          "provinces"
        ]
      }
    },
...
    "/api/v2/provinces/{slug}": {
      "get": {
        "description": "Unified province profile: live oil & gas production summary, mineral project portfolio, and export profile.",
        "operationId": "ProvincesController_detail_v2",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "chubut",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProvinceDetailDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Province slug not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Province detail",
        "tags": [
          "provinces"
        ]
      }
    },
    "/api/v2/provinces/{slug}/production": {
      "get": {
        "description": "Monthly oil & gas production for the province (same shape as operator production). Empty for provinces with no O&G activity. Supports `from` / `to`.",
        "operationId": "ProvincesController_production_v2",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "neuquen",
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive lower bound for date_month (YYYY-MM or YYYY-MM-DD).",
            "schema": {
              "example": "2024-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Inclusive upper bound for date_month (YYYY-MM or YYYY-MM-DD).",
            "schema": {
              "example": "2026-03",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProvinceProductionPointDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Province slug not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Province O&G production time series",
        "tags": [
          "provinces"
        ]
      }
    },
    "/api/v2/provinces/{slug}/exports": {
      "get": {
        "description": "Annual export values for the province, by sector and product.",
        "operationId": "ProvincesController_exports_v2",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "chubut",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProvinceExportRowDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Province slug not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Province exports by sector",
        "tags": [
          "provinces"
        ]
      }
    }
  },
  "info": {
    "title": "Petroldata.ar API",
    "description": "REST API over Argentina Secretaría de Energía oil & gas production data (focus: Vaca Muerta).\n\n## Response envelope\n\nEvery non-GeoJSON endpoint returns:\n\n```json\n{\n  \"data\": ...,\n  \"meta\": {\n    \"source\": \"Secretaría de Energía / datos.energia.gob.ar\",\n    \"dataset\": \"Producción de petróleo y gas por pozo\",\n    \"license\": \"CC-BY-4.0\",\n    \"last_source_update\": \"2026-04-01\",\n    \"last_ingested_at\": \"2026-05-27T14:30:00.000Z\",\n    \"vaca_muerta_filter\": \"formation + unconventional + sub-tipo\"\n  },\n  \"pagination\": { \"page\": 1, \"limit\": 50, \"total\": 1000 }\n}\n```\n\n`pagination` is only present on paginated list endpoints. GeoJSON endpoints (`/geo/*`) return raw GeoJSON.\n\n## Errors\n\n```json\n{ \"error\": { \"code\": \"NOT_FOUND\", \"message\": \"Well not found: 99999\" } }\n```\n\n## Units\n\n- `oil_m3` — cubic metres of oil\n- `oil_bbl` — barrels of oil (1 m³ ≈ 6.2898 bbl)\n- `oil_bbl_d` — barrels per day (oil_bbl ÷ days in month)\n- `gas_thousand_m3` — thousands of cubic metres of gas (Mm³)\n- `gas_mcf` — thousand cubic feet of gas\n- `gas_mmcf_d` — million cubic feet per day\n- `gas_mm3_d` — million m³ per day\n- `boe` — barrels of oil equivalent (oil_bbl + gas_boe)\n\n## Vaca Muerta flag\n\n`vm_combined = formation_vaca_muerta AND unconventional AND sub_resource_type ∈ {Shale, Shale Oil, …}` — pre-computed in the Python pipeline.",
...
  "servers": [
...
  "components": {
    "schemas": {
...
      "MetaDto": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "example": "Secretaría de Energía / datos.energia.gob.ar"
          },
          "dataset": {
            "type": "string",
            "example": "Producción de petróleo y gas por pozo"
          },
          "license": {
            "type": "string",
            "example": "CC-BY-4.0"
          },
          "last_source_update": {
            "type": "object",
            "example": "2026-04-01",
            "nullable": true
          },
          "last_ingested_at": {
            "type": "string",
            "example": "2026-05-27T14:30:00.000Z"
          },
          "vaca_muerta_filter": {
            "type": "string",
            "example": "formation + unconventional + sub-tipo"
          }
        },
        "required": [
          "source",
          "dataset",
          "license",
          "last_source_update",
          "last_ingested_at",
          "vaca_muerta_filter"
        ]
      },
...
      "ApiErrorPayloadDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "NOT_FOUND"
          },
          "message": {
            "type": "string",
            "example": "Well not found: 99999"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorPayloadDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "CompanyListItemDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "ypf"
          },
          "name": {
            "type": "string",
            "example": "YPF S.A."
          },
          "type": {
            "type": "string",
            "example": "oil_and_gas",
            "enum": [
              "mining",
              "oil_and_gas",
              "both"
            ]
          },
          "sector": {
            "type": "string",
            "example": "petroleum"
          },
          "country": {
            "type": "object",
            "example": "Argentina",
            "nullable": true
          },
          "logo_url": {
            "type": "object",
            "example": "https://www.google.com/s2/favicons?domain=www.ypf.com&sz=64",
            "nullable": true
          },
          "stock_ticker": {
            "type": "object",
            "example": "YPF",
            "nullable": true
          },
          "stock_exchange": {
            "type": "object",
            "example": "NYSE",
            "nullable": true
          },
          "is_public": {
            "type": "boolean",
            "example": true
          },
          "project_count_oil_gas": {
            "type": "number",
            "example": 5725
          },
          "project_count_mining": {
            "type": "number",
            "example": 0
          },
          "project_count": {
            "type": "number",
            "example": 5725
          },
          "commodities": {
            "example": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "provinces": {
            "example": [
              "Chubut",
              "Neuquén"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "slug",
          "name",
          "type",
          "sector",
          "country",
          "logo_url",
          "stock_ticker",
          "stock_exchange",
          "is_public",
          "project_count_oil_gas",
          "project_count_mining",
          "project_count",
          "commodities",
          "provinces"
        ]
      },
...
      "CompanyCoordinatesDto": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "object",
            "example": -43.37,
            "nullable": true
          },
          "lng": {
            "type": "object",
            "example": -68.688,
            "nullable": true
          }
        },
        "required": [
          "lat",
          "lng"
        ]
      },
      "CompanyProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Cerro Solo"
          },
          "commodity": {
            "type": "string",
            "example": "Uranio"
          },
          "province": {
            "type": "object",
            "example": "Chubut",
            "nullable": true
          },
          "status": {
            "type": "object",
            "example": "Exploración avanzada",
            "nullable": true
          },
          "coordinates": {
            "$ref": "#/components/schemas/CompanyCoordinatesDto"
          },
          "resources_summary": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            },
            "nullable": true,
            "example": null
          }
        },
        "required": [
          "name",
          "commodity",
          "province",
          "status",
          "coordinates",
          "resources_summary"
        ]
      },
      "TimelineStageDto": {
        "type": "object",
        "properties": {
          "stage": {
            "type": "string",
            "example": "Exploración avanzada"
          },
          "date": {
            "type": "object",
            "example": null,
            "nullable": true
          }
        },
        "required": [
          "stage",
          "date"
        ]
      },
      "OilGasProductionSummaryDto": {
        "type": "object",
        "properties": {
          "oil_production_m3": {
            "type": "object",
            "example": 7327905.65,
            "nullable": true
          },
          "gas_production_m3": {
            "type": "object",
            "example": 3541711.73,
            "nullable": true
          },
          "boe_total": {
            "type": "object",
            "example": 175295959.15,
            "nullable": true
          },
          "well_count": {
            "type": "number",
            "example": 5725
          },
          "provinces": {
            "example": [
              "Chubut",
              "Neuquén"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "oil_production_m3",
          "gas_production_m3",
          "boe_total",
          "well_count",
          "provinces"
        ]
      },
      "CompanyStockDto": {
        "type": "object",
        "properties": {
          "ticker": {
            "type": "string",
            "example": "YPF"
          },
          "price": {
            "type": "object",
            "example": 42.15,
            "nullable": true
          },
          "change_pct": {
            "type": "object",
            "example": 1.2,
            "nullable": true
          },
          "change": {
            "type": "object",
            "example": 0.5,
            "nullable": true
          },
          "currency": {
            "type": "object",
            "example": "USD",
            "nullable": true
          },
          "exchange": {
            "type": "object",
            "example": "NYSE",
            "nullable": true
          },
          "market_time": {
            "type": "object",
            "example": "2026-05-29T20:00:00.000Z",
            "nullable": true
          }
        },
        "required": [
          "ticker",
          "price",
          "change_pct",
          "change",
          "currency",
          "exchange",
          "market_time"
        ]
      },
      "CompanyDetailDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "ypf"
          },
          "name": {
            "type": "string",
            "example": "YPF S.A."
          },
          "type": {
            "type": "string",
            "example": "oil_and_gas"
          },
          "sector": {
            "type": "string",
            "example": "petroleum"
          },
          "country": {
            "type": "object",
            "example": "Argentina",
            "nullable": true
          },
          "website": {
            "type": "object",
            "example": "https://www.ypf.com",
            "nullable": true
          },
          "logo_url": {
            "type": "object",
            "example": "https://www.google.com/s2/favicons?domain=www.ypf.com&sz=64",
            "nullable": true
          },
          "stock_ticker": {
            "type": "object",
            "example": "YPF",
            "nullable": true
          },
          "stock_exchange": {
            "type": "object",
            "example": "NYSE",
            "nullable": true
          },
          "is_public": {
            "type": "boolean",
            "example": true
          },
          "commodities_involved": {
            "example": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "provinces": {
            "example": [
              "Chubut",
              "Neuquén"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "project_count_oil_gas": {
            "type": "number",
            "example": 5725
          },
          "project_count_mining": {
            "type": "number",
            "example": 0
          },
          "projects": {
            "description": "Mineral project portfolio (empty for pure O&G operators).",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyProjectDto"
            }
          },
          "project_timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineStageDto"
            }
          },
          "oil_gas_production_summary": {
            "nullable": true,
            "description": "Present for oil_and_gas / both companies.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OilGasProductionSummaryDto"
              }
            ]
          },
          "stock": {
            "nullable": true,
            "description": "Live quote for public companies, else null.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CompanyStockDto"
              }
            ]
          }
        },
        "required": [
          "slug",
          "name",
          "type",
          "sector",
          "country",
          "website",
          "logo_url",
          "stock_ticker",
          "stock_exchange",
          "is_public",
          "commodities_involved",
          "provinces",
          "project_count_oil_gas",
          "project_count_mining",
          "projects",
          "project_timeline",
          "oil_gas_production_summary",
          "stock"
        ]
      },
      "ProvinceListItemDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "chubut"
          },
          "name": {
            "type": "string",
            "example": "Chubut"
          },
          "iso_code": {
            "type": "object",
            "example": "CT",
            "nullable": true
          },
          "oil_gas_wells": {
            "type": "number",
            "example": 6714
          },
          "mining_projects": {
            "type": "number",
            "example": 8
          },
          "combined_project_count": {
            "type": "number",
            "example": 6722
          },
          "commodities": {
            "example": [
              "Uranio"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "has_oil_gas": {
            "type": "boolean",
            "example": true
          },
          "has_mining": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "slug",
          "name",
          "iso_code",
          "oil_gas_wells",
          "mining_projects",
          "combined_project_count",
          "commodities",
          "has_oil_gas",
          "has_mining"
        ]
      },
...
      "ProvinceTopOperatorDto": {
        "type": "object",
        "properties": {
          "operator_slug": {
            "type": "string",
            "example": "ypf"
          },
          "operator_name": {
            "type": "string",
            "example": "YPF S.A."
          },
          "boe": {
            "type": "number",
            "example": 12345.6
          }
        },
        "required": [
          "operator_slug",
          "operator_name",
          "boe"
        ]
      },
      "ProvinceOilGasDto": {
        "type": "object",
        "properties": {
          "production_oil_bbl_d": {
            "type": "number",
            "example": 245678.9
          },
          "production_gas_mmcf_d": {
            "type": "number",
            "example": 1234.5
          },
          "active_wells": {
            "type": "number",
            "example": 6714
          },
          "vm_pct": {
            "type": "number",
            "example": 0.42,
            "description": "Share of BOE from Vaca Muerta (0–1)."
          },
          "top_operators": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProvinceTopOperatorDto"
            }
          },
          "latest_month": {
            "type": "object",
            "example": "2026-03-01",
            "nullable": true
          }
        },
        "required": [
          "production_oil_bbl_d",
          "production_gas_mmcf_d",
          "active_wells",
          "vm_pct",
          "top_operators",
          "latest_month"
        ]
      },
...
      "ProvinceMiningDto": {
        "type": "object",
        "properties": {
          "project_count": {
            "type": "number",
            "example": 8
          },
          "commodities": {
            "example": [
              "Uranio"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "controllers": {
            "example": [
              "CNEA",
              "Jaguar Uranium Corp."
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProvinceMiningProjectDto"
            }
          }
        },
        "required": [
          "project_count",
          "commodities",
          "controllers",
          "projects"
        ]
      },
      "ProvinceExportRowDto": {
        "type": "object",
        "properties": {
          "sector": {
            "type": "string",
            "example": "petróleo"
          },
          "product": {
            "type": "string",
            "example": "Petróleo crudo"
          },
          "value_annual_usd": {
            "type": "number",
            "example": 781080206
          }
        },
        "required": [
          "sector",
          "product",
          "value_annual_usd"
        ]
      },
      "ProvinceDetailDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Chubut"
          },
          "slug": {
            "type": "string",
            "example": "chubut"
          },
          "iso_code": {
            "type": "object",
            "example": "CT",
            "nullable": true
          },
          "oil_gas": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProvinceOilGasDto"
              }
            ]
          },
          "mining": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProvinceMiningDto"
              }
            ]
          },
          "exports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProvinceExportRowDto"
            }
          },
          "combined_project_count": {
            "type": "number",
            "example": 6722
          }
        },
        "required": [
          "name",
          "slug",
          "iso_code",
          "oil_gas",
          "mining",
          "exports",
          "combined_project_count"
        ]
      },
      "ProvinceProductionPointDto": {
        "type": "object",
        "properties": {
          "date_month": {
            "type": "string",
            "example": "2026-03-01"
          },
          "oil_bbl_d": {
            "type": "number",
            "example": 245678.9
          },
          "gas_mmcf_d": {
            "type": "number",
            "example": 1234.5
          },
          "oil_m3": {
            "type": "number",
            "example": 123456.7
          },
          "gas_thousand_m3": {
            "type": "number",
            "example": 98765.4
          },
          "boe": {
            "type": "number",
            "example": 567890.1
          },
          "active_wells": {
            "type": "number",
            "example": 6714
          }
        },
        "required": [
          "date_month",
          "oil_bbl_d",
          "gas_mmcf_d",
          "oil_m3",
          "gas_thousand_m3",
          "boe",
          "active_wells"
        ]
      }
    }
  }
}